+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - new foods - shroomburger and fruit salad
 - sink supports containers from other mods
 - meat drying
+- fixed crash when is no space in a crafting matrix for a byproduct
 
 ## 0.4.20
 - mushrooms